Improve layout, styling and navigation
I added Typography.js and added site and page nagivation menus. It's
still work in progress: the right (page) navigation needs to be updated
from queries (right now it has static mock content) and the left (site)
navigation needs to be improved - some items are broken and we need to
find a way to specify the order of items.