Fix sizing of center column
Apparently, min-width in grid columns is set to auto by default. This
combined with the presence of pre elements, meant that the columns where
not shrinking with the viewport. Setting min-width to 300px is a
reasonable solution for now. Although the whole grid layout needs to be
made more responsive.