Fix PageToc
This required major refactoring. MDX source files have been migrated to
their own folder outside src/. This seems to work best. It seems that
using a custom template (like page-layout.js) on src/pages/*.mdx files
creates some duplication and conflict. Not sure.

Since a lot of the logic in this site will be moved to a gatsby theme,
it seems this is a step in the right direction.

I also updated dependencies since many of them were lagging behind and
some issues might have been due to that.

I also removed the use of <Todo> since that was causing problems. Will
fix that in future updates.